refactor(headless-docsite): move theme/ into public-docsite-v9-headless
The global tokens.css now lives under apps/public-docsite-v9-headless/theme/
together with the docsite that consumes it. Update import paths in:
  - apps/public-docsite-v9-headless/.storybook/preview.js
  - stories/.storybook/preview.js
  - stories/src/_helpers/withCssModuleSource.ts (?raw inline)
  - project.json build inputs (workspaceRoot -> projectRoot for the docsite)
