+++ apps/public-docsite-v9-headless/project.json
@@ -12,7 +12,7 @@
           "target": "build"
         }
       ],
-      "inputs": ["default", "{workspaceRoot}/.storybook/**", "{projectRoot}/.storybook/**", "{workspaceRoot}/theme/**"]
+      "inputs": ["default", "{workspaceRoot}/.storybook/**", "{projectRoot}/.storybook/**", "{projectRoot}/theme/**"]
     },
     "build-storybook:docsite": {
       "dependsOn": [
@@ -21,7 +21,7 @@
           "target": "build"
         }
       ],
-      "inputs": ["default", "{workspaceRoot}/.storybook/**", "{projectRoot}/.storybook/**", "{workspaceRoot}/theme/**"]
+      "inputs": ["default", "{workspaceRoot}/.storybook/**", "{projectRoot}/.storybook/**", "{projectRoot}/theme/**"]
     }
   }
 }
